membrecias/requirements.txt
amqp==5.1.0
asgiref==3.5.0
attrs==21.4.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.2.3
chardet==4.0.0
click==8.1.1
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cssutils==2.4.0
defusedxml==0.7.1
diff-match-patch==20200713
dj-crontab==0.8.0
Django==2.1.11
django-debug-toolbar==3.2.4
django-excel==0.0.10
django-import-export==2.1.0
django-inlinecss==0.3.0
django-mssql-backend==2.8.1
django-naomi==0.8
django-preventconcurrentlogins==0.8.2
django-pyodbc-azure==2.1.0.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
fontawesomefree==6.0.0
future==0.18.2
importlib==1.0.4
JsonForm==0.0.2
jsonschema==4.4.0
JsonSir==0.0.2
kombu==5.2.4
lml==0.1.0
MarkupPy==1.14
mssql-django==1.1.2
mysqlclient==2.1.0
odfpy==1.4.1
openpyxl==3.0.9
Pillow==9.0.1
prompt-toolkit==3.0.28
pyexcel==0.7.0
pyexcel-io==0.6.6
pyexcel-webio==0.1.4
pynliner==0.8.0
pyodbc==4.0.32
pyrsistent==0.18.1
Python-EasyConfig==0.1.7
pytz==2021.3
PyYAML==6.0
qrcode==7.3.1
Resource==0.2.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
texttable==1.6.4
tzdata==2021.5
vine==5.0.0
waitress==2.1.1
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
